+++ pkg/services/thumbs/thumbimpl/crawler.go
@@ -1,4 +1,4 @@
-package thumbs
+package thumbimpl
 
 import (
 	"context"
@@ -17,6 +17,7 @@ import (
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/live"
 	"github.com/grafana/grafana/pkg/services/rendering"
+	"github.com/grafana/grafana/pkg/services/thumbs"
 )
 
 type simpleCrawler struct {
@@ -26,21 +27,21 @@ type simpleCrawler struct {
 	renderingTimeout time.Duration
 
 	glive                   *live.GrafanaLive
-	thumbnailRepo           thumbnailRepo
-	mode                    CrawlerMode
-	thumbnailKind           ThumbnailKind
-	auth                    CrawlerAuth
+	thumbnailRepo           thumbs.ThumbnailRepo
+	mode                    thumbs.CrawlerMode
+	thumbnailKind           thumbs.ThumbnailKind
+	auth                    thumbs.CrawlerAuth
 	opts                    rendering.Opts
-	status                  crawlStatus
+	status                  thumbs.CrawlStatus
 	statusMutex             sync.RWMutex
-	queue                   []*DashboardWithStaleThumbnail
+	queue                   []*thumbs.DashboardWithStaleThumbnail
 	queueMutex              sync.Mutex
 	log                     log.Logger
 	renderingSessionByOrgId map[int64]rendering.Session
 	dsUidsLookup            getDatasourceUidsForDashboard
 }
 
-func newSimpleCrawler(renderService rendering.Service, gl *live.GrafanaLive, repo thumbnailRepo, cfg *setting.Cfg, settings setting.DashboardPreviewsSettings, dsUidsLookup getDatasourceUidsForDashboard) dashRenderer {
+func newSimpleCrawler(renderService rendering.Service, gl *live.GrafanaLive, repo thumbs.ThumbnailRepo, cfg *setting.Cfg, settings setting.DashboardPreviewsSettings, dsUidsLookup getDatasourceUidsForDashboard) thumbs.DashRenderer {
 	threadCount := int(settings.CrawlThreadCount)
 	c := &simpleCrawler{
 		// temporarily increases the concurrentLimit from the 'cfg.RendererConcurrentRequestLimit' to 'cfg.RendererConcurrentRequestLimit + crawlerThreadCount'
@@ -52,8 +53,8 @@ func newSimpleCrawler(renderService rendering.Service, gl *live.GrafanaLive, rep
 		dsUidsLookup:     dsUidsLookup,
 		thumbnailRepo:    repo,
 		log:              log.New("thumbnails_crawler"),
-		status: crawlStatus{
-			State:    initializing,
+		status: thumbs.CrawlStatus{
+			State:    thumbs.Initializing,
 			Complete: 0,
 			Queue:    0,
 		},
@@ -64,7 +65,7 @@ func newSimpleCrawler(renderService rendering.Service, gl *live.GrafanaLive, rep
 	return c
 }
 
-func (r *simpleCrawler) next(ctx context.Context) (*DashboardWithStaleThumbnail, rendering.Session, rendering.AuthOpts, error) {
+func (r *simpleCrawler) next(ctx context.Context) (*thumbs.DashboardWithStaleThumbnail, rendering.Session, rendering.AuthOpts, error) {
 	r.queueMutex.Lock()
 	defer r.queueMutex.Unlock()
 
@@ -116,13 +117,13 @@ func (r *simpleCrawler) broadcastStatus() {
 	}
 }
 
-type byOrgId []*DashboardWithStaleThumbnail
+type byOrgId []*thumbs.DashboardWithStaleThumbnail
 
 func (d byOrgId) Len() int           { return len(d) }
 func (d byOrgId) Less(i, j int) bool { return d[i].OrgId > d[j].OrgId }
 func (d byOrgId) Swap(i, j int)      { d[i], d[j] = d[j], d[i] }
 
-func (r *simpleCrawler) Run(ctx context.Context, auth CrawlerAuth, mode CrawlerMode, theme models.Theme, thumbnailKind ThumbnailKind) error {
+func (r *simpleCrawler) Run(ctx context.Context, auth thumbs.CrawlerAuth, mode thumbs.CrawlerMode, theme models.Theme, thumbnailKind thumbs.ThumbnailKind) error {
 	res, err := r.renderService.HasCapability(ctx, rendering.ScalingDownImages)
 	if err != nil {
 		return err
@@ -142,7 +143,7 @@ func (r *simpleCrawler) Run(ctx context.Context, auth CrawlerAuth, mode CrawlerM
 		return nil
 	}
 
-	items, err := r.thumbnailRepo.findDashboardsWithStaleThumbnails(ctx, theme, thumbnailKind)
+	items, err := r.thumbnailRepo.FindDashboardsWithStaleThumbnails(ctx, theme, thumbnailKind)
 	if err != nil {
 		r.log.Error("Error when fetching dashboards with stale thumbnails", "err", err.Error())
 		r.queueMutex.Unlock()
@@ -172,9 +173,9 @@ func (r *simpleCrawler) Run(ctx context.Context, auth CrawlerAuth, mode CrawlerM
 
 	r.renderingSessionByOrgId = make(map[int64]rendering.Session)
 	r.queue = items
-	r.status = crawlStatus{
+	r.status = thumbs.CrawlStatus{
 		Started:  runStarted,
-		State:    running,
+		State:    thumbs.Running,
 		Complete: 0,
 	}
 	r.broadcastStatus()
@@ -208,24 +209,24 @@ func (r *simpleCrawler) Run(ctx context.Context, auth CrawlerAuth, mode CrawlerM
 func (r *simpleCrawler) IsRunning() bool {
 	r.statusMutex.Lock()
 	defer r.statusMutex.Unlock()
-	return r.status.State == running
+	return r.status.State == thumbs.Running
 }
 
-func (r *simpleCrawler) Stop() (crawlStatus, error) {
+func (r *simpleCrawler) Stop() (thumbs.CrawlStatus, error) {
 	r.statusMutex.Lock()
-	if r.status.State == running {
-		r.status.State = stopping
+	if r.status.State == thumbs.Running {
+		r.status.State = thumbs.Stopping
 	}
 	r.statusMutex.Unlock()
 
 	return r.Status()
 }
 
-func (r *simpleCrawler) Status() (crawlStatus, error) {
+func (r *simpleCrawler) Status() (thumbs.CrawlStatus, error) {
 	r.statusMutex.RLock()
 	defer r.statusMutex.RUnlock()
 
-	status := crawlStatus{
+	status := thumbs.CrawlStatus{
 		State:    r.status.State,
 		Started:  r.status.Started,
 		Complete: r.status.Complete,
@@ -256,7 +257,7 @@ func (r *simpleCrawler) crawlFinished() {
 	r.statusMutex.Lock()
 	defer r.statusMutex.Unlock()
 
-	r.status.State = stopped
+	r.status.State = thumbs.Stopped
 	r.status.Finished = time.Now()
 }
 
@@ -264,7 +265,7 @@ func (r *simpleCrawler) shouldWalk() bool {
 	r.statusMutex.RLock()
 	defer r.statusMutex.RUnlock()
 
-	return r.status.State == running
+	return r.status.State == thumbs.Running
 }
 
 func (r *simpleCrawler) walk(ctx context.Context, id int) {
@@ -325,7 +326,7 @@ func (r *simpleCrawler) walk(ctx context.Context, id int) {
 					}
 				}()
 
-				thumbnailId, err := r.thumbnailRepo.saveFromFile(ctx, res.FilePath, DashboardThumbnailMeta{
+				thumbnailId, err := r.thumbnailRepo.SaveFromFile(ctx, res.FilePath, thumbs.DashboardThumbnailMeta{
 					DashboardUID: item.Uid,
 					OrgId:        item.OrgId,
 					Theme:        r.opts.Theme,
