+++ pkg/services/thumbs/thumbimpl/datasources_lookup_test.go
@@ -1,4 +1,4 @@
-package thumbs
+package thumbimpl
 
 import (
 	"context"
