+++ pkg/services/thumbs/thumbimpl/testdata/empty_search_response_frame.json
