+++ pkg/services/thumbs/thumbimpl/testdata/search_response_frame.json
